Moving all the entities to an Entity package to ensure encapsulation
