+++ nbproject/private/private.xml
@@ -2,20 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group name="EJEMPLOS PROGRA 2 1-2018">
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/test/domocore/UnitTestSuite.java</file>
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/src/domocore/Home.java</file>
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/test/domocore/HomeIT.java</file>
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/src/domocore/PistonedDoor.java</file>
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/src/domocore/Status.java</file>
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/src/domocore/PistonedWindow.java</file>
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/src/domocore/Domocore.java</file>
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/src/domocore/Openable.java</file>
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/src/domocore/Piston.java</file>
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/src/domocore/OpenCloseElement.java</file>
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/src/domocore/Actuator.java</file>
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/src/domocore/OpenCloseElementByPiston.java</file>
-            <file>file:/C:/Users/Jheron%20Chacon/Documents/NetBeansProjects/GestorAcademico/GestorAcademico/Domocore/src/util/Console.java</file>
-        </group>
+        <group name="EJEMPLOS PROGRA 2 1-2018"/>
     </open-files>
 </project-private>
