+++ src/Entities/PistonedDoor.java
@@ -3,15 +3,16 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package domocore;
+package Entities;
 
 /**
  *
  * @author Jheron Chacon
  */
-public interface Openable {
-    public void open();
-    public void open(int percentage);
-    public void close();
-    public void close(int percentage);
+public class PistonedDoor extends OpenCloseElementByPiston{
+    
+    public PistonedDoor(int id){
+        super(id);
+    }
+    
 }
